chore(related-folder-restructure): Implements a folder restrucutre to hold related files
[Finishes #163724518]
